move sound to soundStore
