+++ src/components/game-setup/VolumeButton.tsx
@@ -4,11 +4,11 @@ import { useMainStore } from "../../hooks/useMainStore";
 import styles from "./VolumeButton.module.scss";
 
 export const VolumeButton = observer(() => {
-  const { soundVolume, toggleSound } = useMainStore();
+  const { sound } = useMainStore();
 
   return (
-    <button type="button" className={styles.sound} onClick={toggleSound}>
-      {soundVolume === "1" ? (
+    <button type="button" className={styles.sound} onClick={sound.toggleSound}>
+      {sound.soundVolume === "1" ? (
         <span className="material-symbols-outlined">volume_up</span>
       ) : (
         <span className="material-symbols-outlined">volume_off</span>
